requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.14.0
openai==2.32.0
packaging==26.1
pluggy==1.6.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
